Moved exports out of index
